+++ src/devices/air.ts
@@ -92,7 +92,7 @@ export class airSensor {
 		try{
 			currentValue = batteryStatus.getCharacteristic(this.platform.Characteristic.StatusLowBattery).value;
 			if (currentValue === 1) {
-				this.platform.log.warn('Battery Status Low');
+				this.platform.log.warn('Air Sensor Battery Status Low');
 			}
 		}catch (error) {
 			this.platform.log.error('caught low battery error');